airflow.requirements.txt
# Add any additional Python dependencies below:
bigeye-sdk==0.4.54
bigeye-cli==0.3.32
bigeye-airflow==0.1.20
dbt-core==1.4.5
dbt-snowflake==1.4.2
pyOpenSSL==22.0.0
snowflake-connector-python==3.0.4
snowflake-sqlalchemy==1.4.7
apache-airflow-providers-snowflake==4.1.0